Mock pyarrow when restoring a snapshot
